moved semver modes to a modes package - renamed the mode api - removed mode detector - added delimiters to git modes - moved mode detection to git modes - implemented delimiter logic for git modes
