+++ pkg/modes/minor_test.go
@@ -1,4 +1,4 @@
-package semver
+package modes
 
 import (
 	"testing"
@@ -6,16 +6,16 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestPatchMode_PatchConstant(t *testing.T) {
+func TestMinorMode_MinorConstant(t *testing.T) {
 	t.Run("CheckConstant", func(t *testing.T) {
-		var want = "patch"
-		var got = Patch
+		var want = "minor"
+		var got = Minor
 
 		assert.Equal(t, want, got, `want: "%s", got: "%s"`, want, got)
 	})
 }
 
-func TestPatchMode_Increment(t *testing.T) {
+func TestMinorMode_Increment(t *testing.T) {
 	type Test struct {
 		Name          string
 		TargetVersion string
@@ -23,15 +23,16 @@ func TestPatchMode_Increment(t *testing.T) {
 	}
 
 	var tests = []Test{
-		{Name: "HappyPath", TargetVersion: "0.0.1", Want: "0.0.2"},
-		{Name: "DiscardPreBuild", TargetVersion: "0.0.8-pre+001", Want: "0.0.9"},
-		{Name: "NoResetMajorMinor", TargetVersion: "5.4.3", Want: "5.4.4"},
+		{Name: "HappyPath", TargetVersion: "0.1.0", Want: "0.2.0"},
+		{Name: "ResetPatch", TargetVersion: "0.2.3", Want: "0.3.0"},
+		{Name: "NoResetMajor", TargetVersion: "6.7.0", Want: "6.8.0"},
+		{Name: "DiscardPreBuild", TargetVersion: "0.6.0-pre+001", Want: "0.7.0"},
 	}
 
 	for _, test := range tests {
 		t.Run(test.Name, func(t *testing.T) {
 			var want = test.Want
-			var mode = NewPatchMode()
+			var mode = NewMinorMode()
 			var got, err = mode.Increment(test.TargetVersion)
 
 			assert.NoError(t, err)
@@ -51,7 +52,7 @@ func TestPatchMode_Increment(t *testing.T) {
 
 	for _, test := range errorTests {
 		t.Run(test.Name, func(t *testing.T) {
-			var mode = NewPatchMode()
+			var mode = NewMinorMode()
 			var _, got = mode.Increment(test.TargetVersion)
 			assert.Error(t, got)
 		})
